+++ src/main/java/io/reactivesocket/UnlimitedLeaseGovernor.java
@@ -5,7 +5,7 @@ import io.reactivesocket.internal.Responder;
 public class UnlimitedLeaseGovernor implements LeaseGovernor {
     @Override
     public void register(Responder responder) {
-        responder.sendLease(Long.MAX_VALUE, Long.MAX_VALUE);
+        responder.sendLease(Integer.MAX_VALUE, Integer.MAX_VALUE);
     }
 
     @Override
